Migrate the legacy ConnectionUnitTest
