+++ Code/Tools/AssetProcessor/assetprocessor_test_files.cmake
@@ -80,7 +80,6 @@ set(FILES
     native/unittests/AssetScannerUnitTests.cpp
     native/unittests/AssetScannerUnitTests.h
     native/unittests/ConnectionUnitTests.cpp
-    native/unittests/ConnectionUnitTests.h
     native/unittests/ConnectionManagerUnitTests.cpp
     native/unittests/FileWatcherUnitTests.cpp
     native/unittests/FileWatcherUnitTests.h
